feat: api operations into di framework
