+++ src/next.web.core/services/JsAuthenicateHandler.cs
@@ -44,7 +44,7 @@ namespace next.web.core.services
             }
         }
 
-        public async virtual Task<FormSubmissionResponse> Submit(FormSubmissionModel model, ISession session)
+        public async virtual Task<FormSubmissionResponse> Submit(FormSubmissionModel model, ISession session, IApiWrapper? wrapper = null)
         {
             var response = FormResponses.GetDefault(model.FormName) ?? new();
             try
