+++ src/next.web/ProgramExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using legallead.desktop.interfaces;
 using Microsoft.AspNetCore.Rewrite;
-using next.web.Controllers;
 using next.web.core.util;
 using next.web.Services;
 using System.Diagnostics.CodeAnalysis;
